Flatten models. Rename calc_type -> calctype. Refactor modules for simplicity.
