+++ qcio/models/__init__.py
@@ -1,4 +1,4 @@
 from .base_io import *  # noqa: F403
-from .file import *  # noqa: F403
+from .file_io import *  # noqa: F403
 from .molecule import *  # noqa: F403
-from .single_point import *  # noqa: F403
+from .single_point_io import *  # noqa: F403
