+++ tests/conftest.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import numpy as np
 import pytest
 
-from qcio import SinglePointInput, SinglePointSuccessfulOutput
+from qcio import SinglePointInput, SinglePointResult
 from qcio.utils import water as water_mol
 
 
@@ -24,19 +24,17 @@ def sp_input(water):
     """Create a function that returns a SinglePointInput object with a specified
     calculation type."""
 
-    def _create_sp_input(calc_type):
+    def _create_sp_input(calctype):
         return SinglePointInput(
             molecule=water,
-            program_args={
-                "calc_type": calc_type,
-                "model": {"method": "hf", "basis": "sto-3g"},
-                "keywords": {
-                    "maxiter": 100,
-                    "purify": "no",
-                    "some-bool": False,
-                    "displacement": 1e-3,
-                    "thermo_temp": 298.15,
-                },
+            calctype=calctype,
+            model={"method": "hf", "basis": "sto-3g"},
+            keywords={
+                "maxiter": 100,
+                "purify": "no",
+                "some-bool": False,
+                "displacement": 1e-3,
+                "thermo_temp": 298.15,
             },
         )
 
@@ -44,7 +42,7 @@ def sp_input(water):
 
 
 @pytest.fixture
-def sp_successful_output(sp_input):
+def sp_result(sp_input):
     """Create SinglePointOutput object"""
     sp_inp_energy = sp_input("energy")
     energy = 1.0
@@ -52,7 +50,7 @@ def sp_successful_output(sp_input):
     gradient = np.arange(n_atoms * 3).reshape(n_atoms, 3)
     hessian = np.arange(n_atoms**2 * 3**2).reshape(n_atoms * 3, n_atoms * 3)
 
-    return SinglePointSuccessfulOutput(
+    return SinglePointResult(
         input_data=sp_inp_energy,
         stdout="program standard out...",
         computed={
