+++ tests/test_single_point.py
@@ -3,9 +3,9 @@ from qcelemental.models import AtomicInput, AtomicResult, FailedOperation
 
 from qcio import (
     Provenance,
-    SinglePointComputedProperties,
-    SinglePointFailedOutput,
-    SinglePointSuccessfulOutput,
+    SinglePointComputedProps,
+    SinglePointFailure,
+    SinglePointResult,
     Wavefunction,
 )
 
@@ -13,17 +13,17 @@ from qcio import (
 def test_computed_gradient_converted_np_array():
     """Test that SinglePointResult converts gradient to np array"""
     gradient = [1, 2, 3, 4, 5, 6]
-    result = SinglePointComputedProperties(gradient=gradient)
-    assert isinstance(result.gradient, np.ndarray)
-    assert result.gradient.dtype == np.float64
+    computed = SinglePointComputedProps(gradient=gradient)
+    assert isinstance(computed.gradient, np.ndarray)
+    assert computed.gradient.dtype == np.float64
 
 
 def test_computed_hessian_converted_np_array():
     """Test that SinglePointResult converts gradient to np array"""
     hessian = [1, 2, 3, 4, 5, 6, 7, 8, 9]
-    result = SinglePointComputedProperties(hessian=hessian)
-    assert isinstance(result.hessian, np.ndarray)
-    assert result.hessian.dtype == np.float64
+    computed = SinglePointComputedProps(hessian=hessian)
+    assert isinstance(computed.hessian, np.ndarray)
+    assert computed.hessian.dtype == np.float64
 
 
 def test_single_point_success_casts_gradient_to_n_by_3(sp_input):
@@ -31,16 +31,16 @@ def test_single_point_success_casts_gradient_to_n_by_3(sp_input):
     sp_inp_energy = sp_input("energy")
     n_atoms = len(sp_inp_energy.molecule.symbols)
     gradient = [i for i in range(n_atoms * 3)]
-    success = SinglePointSuccessfulOutput(
+    result = SinglePointResult(
         input_data=sp_inp_energy,
         computed={
             "gradient": gradient,
         },
         provenance={"program": "qcio-test-suite"},
     )
-    assert success.computed.gradient.shape == (n_atoms, 3)
+    assert result.computed.gradient.shape == (n_atoms, 3)
     # Assert that gradient is numpy array of floats
-    assert success.computed.gradient.dtype == np.float64
+    assert result.computed.gradient.dtype == np.float64
 
 
 def test_single_point_success_casts_hessian_to_3n_by_3n(sp_input):
@@ -49,16 +49,16 @@ def test_single_point_success_casts_hessian_to_3n_by_3n(sp_input):
     n_atoms = len(sp_inp_energy.molecule.symbols)
     hessian = [i for i in range(n_atoms**2 * 3**2)]
 
-    output = SinglePointSuccessfulOutput(
+    result = SinglePointResult(
         input_data=sp_inp_energy,
         computed={
             "hessian": hessian,
         },
         provenance={"program": "qcio-test-suite"},
     )
-    assert output.computed.hessian.shape == (n_atoms * 3, n_atoms * 3)
+    assert result.computed.hessian.shape == (n_atoms * 3, n_atoms * 3)
     # Assert that hessian is numpy array of floats
-    assert output.computed.hessian.dtype == np.float64
+    assert result.computed.hessian.dtype == np.float64
 
 
 def test_return_result(sp_input):
@@ -69,7 +69,7 @@ def test_return_result(sp_input):
     gradient = np.arange(n_atoms * 3).reshape(n_atoms, 3)
     hessian = np.arange(n_atoms**2 * 3**2).reshape(n_atoms * 3, n_atoms * 3)
 
-    output = SinglePointSuccessfulOutput(
+    result = SinglePointResult(
         input_data=sp_inp_energy,
         computed={
             "energy": energy,
@@ -78,45 +78,37 @@ def test_return_result(sp_input):
         },
         provenance={"program": "qcio-test-suite"},
     )
-    assert output.return_result == energy
-    assert output.return_result == output.computed.energy
+    assert result.return_result == energy
+    assert result.return_result == result.computed.energy
 
     sp_inp_grad = sp_input("gradient")
-    output = SinglePointSuccessfulOutput(
-        **{**output.dict(), **{"input_data": sp_inp_grad}}
-    )
+    result = SinglePointResult(**{**result.dict(), **{"input_data": sp_inp_grad}})
 
-    assert np.array_equal(output.return_result, gradient)
-    assert np.array_equal(output.return_result, output.computed.gradient)
+    assert np.array_equal(result.return_result, gradient)
+    assert np.array_equal(result.return_result, result.computed.gradient)
 
     sp_inp_hessian = sp_input("hessian")
-    output = SinglePointSuccessfulOutput(
-        **{**output.dict(), **{"input_data": sp_inp_hessian}}
-    )
+    result = SinglePointResult(**{**result.dict(), **{"input_data": sp_inp_hessian}})
 
-    assert np.array_equal(output.return_result, hessian)
-    assert np.array_equal(output.return_result, output.computed.hessian)
+    assert np.array_equal(result.return_result, hessian)
+    assert np.array_equal(result.return_result, result.computed.hessian)
 
 
-def test_successful_output_serialization(sp_successful_output):
+def test_successful_output_serialization(sp_result):
     """Test that successful result serializes and deserializes"""
-    serialized = sp_successful_output.json()
+    serialized = sp_result.json()
     # model_validate_json in pydantic v2
-    deserialized = SinglePointSuccessfulOutput.parse_raw(serialized)
-    assert deserialized == sp_successful_output
-    assert deserialized.computed == sp_successful_output.computed
-    assert deserialized.input_data == sp_successful_output.input_data
+    deserialized = SinglePointResult.parse_raw(serialized)
+    assert deserialized == sp_result
+    assert deserialized.computed == sp_result.computed
+    assert deserialized.input_data == sp_result.input_data
     assert deserialized.provenance.program == "qcio-test-suite"
-    assert deserialized.stdout == sp_successful_output.stdout
-    assert deserialized.extras == sp_successful_output.extras
-    assert deserialized.return_result == sp_successful_output.return_result
-    assert deserialized.return_result == sp_successful_output.computed.energy
-    assert np.array_equal(
-        deserialized.computed.gradient, sp_successful_output.computed.gradient
-    )
-    assert np.array_equal(
-        deserialized.computed.hessian, sp_successful_output.computed.hessian
-    )
+    assert deserialized.stdout == sp_result.stdout
+    assert deserialized.extras == sp_result.extras
+    assert deserialized.return_result == sp_result.return_result
+    assert deserialized.return_result == sp_result.computed.energy
+    assert np.array_equal(deserialized.computed.gradient, sp_result.computed.gradient)
+    assert np.array_equal(deserialized.computed.hessian, sp_result.computed.hessian)
 
 
 def test_to_failure(sp_input):
@@ -158,16 +150,18 @@ def test_input_to_qcel(sp_input):
         qcel_atomic_input.molecule.geometry, sp_inp_energy.molecule.geometry
     )
     assert qcel_atomic_input.molecule.molecular_charge == sp_inp_energy.molecule.charge
-    assert qcel_atomic_input.model.method == sp_inp_energy.program_args.model.method
-    assert qcel_atomic_input.model.basis == sp_inp_energy.program_args.model.basis
+    assert qcel_atomic_input.model.method == sp_inp_energy.model.method
+    assert qcel_atomic_input.model.basis == sp_inp_energy.model.basis
     assert qcel_atomic_input.driver == "energy"
-    assert qcel_atomic_input.keywords == sp_inp_energy.program_args.keywords
+    assert qcel_atomic_input.keywords == sp_inp_energy.keywords
     assert qcel_atomic_input.extras == sp_inp_energy.extras
 
 
 def test_input_to_output_from_qcel_output_with_wfn(sp_input):
     """Test that input converts to qcel input"""
     sp_inp_energy = sp_input("energy")
+
+    # Check that the input converts to a qcel input
     qcel_atomic_input = AtomicInput(**sp_inp_energy.to_qcel())
 
     qcel_atomic_output = AtomicResult(
@@ -194,17 +188,16 @@ def test_input_to_output_from_qcel_output_with_wfn(sp_input):
             },
         },
     )
-    sp_successful_output = sp_inp_energy.to_output_from_qcel(qcel_atomic_output.dict())
+    result = sp_inp_energy.to_output_from_qcel(qcel_atomic_output.dict())
 
-    assert sp_successful_output.input_data == sp_inp_energy
-    assert sp_successful_output.computed.energy == qcel_atomic_output.return_result
+    assert result.input_data == sp_inp_energy
+    assert result.computed.energy == qcel_atomic_output.return_result
     assert (
-        sp_successful_output.computed.nuclear_repulsion_energy
+        result.computed.nuclear_repulsion_energy
         == qcel_atomic_output.properties.nuclear_repulsion_energy
     )
     assert (
-        sp_successful_output.computed.calcinfo_natoms
-        == qcel_atomic_output.properties.calcinfo_natom
+        result.computed.calcinfo_natoms == qcel_atomic_output.properties.calcinfo_natom
     )
 
 
@@ -227,17 +220,16 @@ def test_input_to_output_from_qcel_output_no_wfn(sp_input):
             },
         },
     )
-    sp_successful_output = sp_inp_energy.to_output_from_qcel(qcel_atomic_output.dict())
+    result = sp_inp_energy.to_output_from_qcel(qcel_atomic_output.dict())
 
-    assert sp_successful_output.input_data == sp_inp_energy
-    assert sp_successful_output.computed.energy == qcel_atomic_output.return_result
+    assert result.input_data == sp_inp_energy
+    assert result.computed.energy == qcel_atomic_output.return_result
     assert (
-        sp_successful_output.computed.nuclear_repulsion_energy
+        result.computed.nuclear_repulsion_energy
         == qcel_atomic_output.properties.nuclear_repulsion_energy
     )
     assert (
-        sp_successful_output.computed.calcinfo_natoms
-        == qcel_atomic_output.properties.calcinfo_natom
+        result.computed.calcinfo_natoms == qcel_atomic_output.properties.calcinfo_natom
     )
 
 
@@ -261,8 +253,8 @@ def test_input_to_output_from_qcel_failed_operation(sp_input):
 
     sp_inp_energy = sp_input("energy")
 
-    sp_failed_output = sp_inp_energy.to_output_from_qcel(failed_operation.dict())
-    assert isinstance(sp_failed_output, SinglePointFailedOutput)
-    assert sp_failed_output.input_data == sp_inp_energy
-    assert sp_failed_output.traceback == failed_operation.error.error_message
-    assert sp_failed_output.extras == failed_operation.extras
+    failure = sp_inp_energy.to_output_from_qcel(failed_operation.dict())
+    assert isinstance(failure, SinglePointFailure)
+    assert failure.input_data == sp_inp_energy
+    assert failure.traceback == failed_operation.error.error_message
+    assert failure.extras == failed_operation.extras
